requirements.txt
Pillow==10.0.0
PyGetWindow==0.0.9
pyglet==2.0.9
